cloud/getImages/requirements.txt
functions-framework==3.*
google-cloud-storage
Flask